+++ vinted/__init__.py
@@ -11,7 +11,7 @@ from vinted.exceptions import (
 )
 from vinted.models.item import CatalogItem, DetailedItem
 
-__version__ = "1.0.0"
+__version__ = "1.0.1"
 
 __all__ = [
     "VintedClient",